requirements.txt
mysql~=0.0.3
mysql-connector-python~=8.3.0
flask~=3.0.2
itsdangerous~=2.1.2
PyJWT~=2.8.0
python-dotenv~=1.0.1
werkzeug~=3.0.1
requests~=2.31.0